+++ springboot/src/main/java/com/springboot/project/model/UserMessageModel.java
@@ -25,7 +25,7 @@ public class UserMessageModel {
 
     private Long totalPage;
 
-    private Integer pageNum;
+    private Long pageNum;
 
     private String url;
 }
