+++ springboot/src/test/java/com/springboot/project/test/service/StorageSpaceService/StorageSpaceServiceDeleteStorageSpaceEntityTest.java
@@ -15,11 +15,11 @@ public class StorageSpaceServiceDeleteStorageSpaceEntityTest extends BaseTest {
     @Test
     public void test() {
         this.storageSpaceService.deleteStorageSpaceEntity(folderName);
-        var totalPage = this.storageSpaceService.getStorageSpaceListByPagination(1, 1).getTotalPage();
+        var totalPage = this.storageSpaceService.getStorageSpaceListByPagination(1L, 1L).getTotalPage();
         var list = Observable.interval(0, TimeUnit.SECONDS).take(totalPage).concatMap((s) -> {
-            int pageNum = s.intValue() + 1;
+            Long pageNum = s + 1;
             return Observable
-                    .fromIterable(this.storageSpaceService.getStorageSpaceListByPagination(pageNum, 1).getList());
+                    .fromIterable(this.storageSpaceService.getStorageSpaceListByPagination(pageNum, 1L).getList());
         }).toList().blockingGet();
         assertTrue(JinqStream.from(list).where(s -> s.getFolderName().equals(folderName)).exists());
     }
